- minor changes so the FreeBSD changes will work under Linux
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/trunk@1145 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/act.wizard.c
@@ -1618,10 +1618,10 @@ void do_stat(struct char_data *ch, char *argument, int cmd)
                           classes[k->specials.remortclass - 1].name, color1);
             }
 
-            sprintf(buf, "%sBirth : [%s%d%s]secs, Logon[%s%d%s]secs, "
+            sprintf(buf, "%sBirth : [%s%ld%s]secs, Logon[%s%ld%s]secs, "
                          "Played[%s%d%s]secs",
-                    color1, color2, k->player.time.birth, color1, color2,
-                    k->player.time.logon, color1, color2,
+                    color1, color2, (long)k->player.time.birth, color1, color2,
+                    (long)k->player.time.logon, color1, color2,
                     k->player.time.played, color1);
             act(buf, FALSE, ch, 0, 0, TO_CHAR);
